refactor: reduce public footprint, reorganize packages (#4091)
* reduce public footprint, reorganize packages

* reorganize packages

* reorganize

* move long distance freight generation to matsim-germany
