--- contribs/application/src/main/java/org/matsim/application/prepare/freight/tripGeneration/DefaultDepartureTimeCalculator.java
@@ -1,12 +1,0 @@
-package org.matsim.application.prepare.freight.tripGeneration;
-
-import java.util.Random;
-
-public class DefaultDepartureTimeCalculator implements FreightAgentGenerator.DepartureTimeCalculator {
-    private final Random rnd = new Random(1111);
-
-    @Override
-    public double getDepartureTime() {
-        return rnd.nextInt(24) * 3600.0;
-    }
-}
