+++ contribs/small-scale-traffic-generation/src/main/java/org/matsim/commercialDemandGenerationBasic/DemandGenerationSpecificationForParcelDelivery.java
@@ -1,4 +1,4 @@
-package org.matsim.freightDemandGeneration;
+package org.matsim.commercialDemandGenerationBasic;
 
 import com.google.common.base.Joiner;
 import org.apache.commons.math3.distribution.EnumeratedDistribution;
@@ -19,15 +19,15 @@ import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
 
-import static org.matsim.freightDemandGeneration.DemandReaderFromCSV.getHomeCoord;
+import static org.matsim.commercialDemandGenerationBasic.DemandReaderFromCSV.getHomeCoord;
 
 /**
  * This class can be used to generate demand for parcel delivery shipments by using the tool FreightDemandGeneration.
  *
- * @see FreightDemandGeneration
+ * @see BasicCommercialDemandGeneration
  * @Author Ricardo Ewert (based on the work of the master thesis of Ana Ueberhorst)
  */
-public class DemandGenerationSpecificationForParcelDelivery extends DefaultDemandGenerationSpecification {
+class DemandGenerationSpecificationForParcelDelivery extends DefaultDemandGenerationSpecification {
 	//	private static double PACKAGES_PER_PERSON = 0.0686; //default value
 	private static double PACKAGES_PER_PERSON = 0.14; //default value
